requirements.txt
tqdm
pdf2image
pytesseract
Pillow
python-docx